refactor: improve import patterns using barrel exports
Refactor import statements across the codebase to utilize the newly created
barrel exports for cleaner and more maintainable code organization.

## Changes Made

### Hook Import Improvements
- **useExportHandlers**: Use `@features/schema` barrel export
- **useTerminalCalculations**: Consolidate store hooks and feature imports
- **useSearchHandlers**: Combine navigation and search hooks from `@store/hooks`

### Component Import Standardization
- **JsonViewer**: Use `@features/json-rendering` barrel export
- **SchemaViewer**: Use `@features/schema` barrel export

### Benefits

#### Before ❌
```typescript
import { exportToFile } from "@features/schema/utils/fileExport";
import { useDebugInfo } from "@store/hooks/useDebug";
import { useJqState } from "@store/hooks/useJq";
import { jsonHighlighter } from "@features/json-rendering/utils/jsonHighlighter";
```

#### After ✅
```typescript
import { exportToFile } from "@features/schema";
import { useDebugInfo, useJqState } from "@store/hooks";
import { jsonHighlighter } from "@features/json-rendering";
```

### Key Improvements
- **Reduced Import Paths**: Shorter, more intuitive import statements
- **Consolidated Imports**: Multiple related imports combined into single statements
- **Consistent Patterns**: All imports follow the same barrel export pattern
- **Better Maintainability**: Easier refactoring and code organization

## Verification

- ✅ TypeScript compilation passes
- ✅ Build process completes successfully
- ✅ All existing functionality preserved
- ✅ No breaking changes to public APIs

This builds upon the comprehensive barrel export infrastructure added in the
previous code structure improvements, making the codebase more developer-friendly
and maintainable.
